requirements_diffusion.txt
anyio==4.12.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
fonttools==4.61.0
fqdn==1.5.1
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.7.0
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
kiwisolver==1.4.9
lark==1.3.1
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
narwhals==2.14.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.0
notebook-shim==0.2.4

# === 핵심 수치 안정 조합 ===
numpy==1.26.4
opencv-python==4.9.0.80

overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
polars==1.36.1
polars-runtime-32==1.36.1
prometheus-client==0.23.1
prompt-toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.23
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.3
send2trash==1.8.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0

tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.236
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.6.2
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0

# === Diffusion / HF ===
gradio
segment-anything==1.0

diffusers==0.35.1
transformers==4.48.3
huggingface-hub>=0.34.0,<0.37.0
tokenizers>=0.21,<0.22

safetensors>=0.4.3
accelerate>=0.27.2
soundfile